+++ Drop-In/src/main/java/com/braintreepayments/api/dropin/DropInRequest.java
@@ -28,6 +28,7 @@ public class DropInRequest implements Parcelable {
 
     private GooglePaymentRequest mGooglePaymentRequest;
     private PayPalRequest mPayPalRequest;
+
     private boolean mGooglePaymentEnabled = true;
     private boolean mMaskCardNumber = false;
     private boolean mMaskSecurityCode = false;
@@ -37,6 +38,8 @@ public class DropInRequest implements Parcelable {
     private boolean mCardEnabled = true;
     private boolean mDefaultVaultValue = true;
     private boolean mShowCheckBoxToAllowVaultOverride = false;
+    private boolean mVaultVenmo = false;
+
     private int mCardholderNameStatus = CardForm.FIELD_DISABLED;
 
     public DropInRequest() {}
@@ -215,6 +218,16 @@ public class DropInRequest implements Parcelable {
     }
 
     /**
+     * @param defaultValue the default value used to determine if Drop-in should vault the customer's venmo payment method. Must be set to `false` when using a client token without a `customerId`.
+     *
+     * This value is {@code false} by default.
+     */
+    public DropInRequest vaultVenmo(boolean defaultValue) {
+        mVaultVenmo = defaultValue;
+        return this;
+    };
+
+    /**
      * @param customerCheckBoxEnabled {@code true} shows save card CheckBox to allow user to choose whether or not to vault their card.
      * {@code false} does not show Save Card CheckBox.
      */
@@ -295,6 +308,8 @@ public class DropInRequest implements Parcelable {
         return mMaskSecurityCode;
     }
 
+    boolean shouldVaultVenmo() { return mVaultVenmo; }
+
     boolean isVaultManagerEnabled() {
         return mVaultManagerEnabled;
     }
@@ -335,6 +350,7 @@ public class DropInRequest implements Parcelable {
         dest.writeInt(mCardholderNameStatus);
         dest.writeByte(mDefaultVaultValue ? (byte) 1 : (byte) 0);
         dest.writeByte(mShowCheckBoxToAllowVaultOverride ? (byte) 1 : (byte) 0);
+        dest.writeByte(mVaultVenmo ? (byte) 1 : (byte) 0);
     }
 
     protected DropInRequest(Parcel in) {
@@ -355,6 +371,7 @@ public class DropInRequest implements Parcelable {
         mCardholderNameStatus = in.readInt();
         mDefaultVaultValue = in.readByte() != 0;
         mShowCheckBoxToAllowVaultOverride = in.readByte() != 0;
+        mVaultVenmo = in.readByte() != 0;
     }
 
     public static final Creator<DropInRequest> CREATOR = new Creator<DropInRequest>() {